Move utility functions around in desperate attempt to resolve import cycles
