+++ pkg/util/network.go
@@ -4,15 +4,12 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"net"
 	"net/http"
 	"os"
 	"path/filepath"
-	"syscall"
 	"time"
 
 	"github.com/cheggaaa/pb"
-	"github.com/drud/ddev/pkg/dockerutil"
 	"github.com/drud/ddev/pkg/output"
 	log "github.com/sirupsen/logrus"
 )
@@ -126,36 +123,4 @@ func EnsureHTTPStatus(o *HTTPOptions) error {
 
 	}
 	return fmt.Errorf("failed to match status code: %d: %v", o.ExpectedStatus, err)
-}
-
-// IsPortActive checks to see if the given port on docker IP is answering.
-func IsPortActive(port string) bool {
-	dockerIP, err := dockerutil.GetDockerIP()
-	if err != nil {
-		Warning("Failed to get docker IP address: %v", err)
-		return false
-	}
-
-	conn, err := net.Dial("tcp", dockerIP+":"+port)
-
-	// If we were able to connect, something is listening on the port.
-	if err == nil {
-		_ = conn.Close()
-		return true
-	}
-	// If we get ECONNREFUSED the port is not active.
-	oe, ok := err.(*net.OpError)
-	if ok {
-		syscallErr, ok := oe.Err.(*os.SyscallError)
-
-		// On Windows, WSAECONNREFUSED (10061) results instead of ECONNREFUSED. And golang doesn't seem to have it.
-		var WSAECONNREFUSED syscall.Errno = 10061
-
-		if ok && (syscallErr.Err == syscall.ECONNREFUSED || syscallErr.Err == WSAECONNREFUSED) {
-			return false
-		}
-	}
-	// Otherwise, hmm, something else happened. It's not a fatal or anything.
-	Warning("Unable to properly check port status: %v", oe)
-	return false
 }
