+++ pkg/version/version.go
@@ -2,8 +2,9 @@ package version
 
 import (
 	"fmt"
-	"github.com/drud/ddev/pkg/dockerutil"
-	"github.com/drud/ddev/pkg/util"
+	"github.com/fsouza/go-dockerclient"
+	"os/exec"
+	"strings"
 )
 
 // MariaDBDefaultVersion is the default version we use in the db container
@@ -98,10 +99,10 @@ func GetVersionInfo() map[string]string {
 	versionInfo["commit"] = COMMIT
 	versionInfo["domain"] = DDevTLD
 	versionInfo["build info"] = BUILDINFO
-	if versionInfo["docker"], err = util.GetDockerVersion(); err != nil {
+	if versionInfo["docker"], err = GetDockerVersion(); err != nil {
 		_ = fmt.Errorf("failed to GetDockerVersion(): %v", err)
 	}
-	if versionInfo["docker-compose"], err = dockerutil.GetDockerComposeVersion(); err != nil {
+	if versionInfo["docker-compose"], err = GetDockerComposeVersion(); err != nil {
 		_ = fmt.Errorf("failed to GetDockerComposeVersion(): %v", err)
 	}
 
@@ -130,4 +131,46 @@ func GetDBAImage() string {
 // GetDBAImage returns the correctly formatted dba image:tag reference
 func GetBgsyncImage() string {
 	return fmt.Sprintf("%s:%s", BgsyncImg, BgsyncTag)
+}
+
+// GetDockerComposeVersion runs docker-compose -v to get the current version
+func GetDockerComposeVersion() (string, error) {
+	if DockerComposeVersion != "" {
+		return DockerComposeVersion, nil
+	}
+
+	executableName := "docker-compose"
+
+	path, err := exec.LookPath(executableName)
+	if err != nil {
+		return "", fmt.Errorf("no docker-compose")
+	}
+
+	out, err := exec.Command(path, "version", "--short").Output()
+	if err != nil {
+		return "", err
+	}
+
+	v := string(out)
+	DockerComposeVersion = strings.TrimSpace(v)
+	return DockerComposeVersion, nil
+}
+
+// GetDockerVersion gets the cached or api-sourced version of docker engine
+func GetDockerVersion() (string, error) {
+	if DockerVersion != "" {
+		return DockerVersion, nil
+	}
+	var client *docker.Client
+	var err error
+	if client, err = docker.NewClientFromEnv(); err != nil {
+		return "", err
+	}
+
+	v, err := client.Version()
+	if err != nil {
+		return "", err
+	}
+	DockerVersion = v.Get("Version")
+	return DockerVersion, nil
 }
